pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Flask app to analyze pages"
authors = ["Egor Gerimovich <cherund@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
validators = "^0.28.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
